components/jaeger-collector: move to memory store
This commit reconfigures Jaeger to use its in-memory store rather than
Badger, due to performance and operational issues we have been having in
production. Note that unlike Badger, the in-memory store does not
support time-based retention; instead it only supports limiting the
total number of traces. This limit is parameterized as an OpenShift
template variable.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>