components,environments: obs RBAC + tenancy
This commit enables multi-tenancy in the configuration. We modify the
dev environment and tests to deploy a Dex and configure the
Observatorium API to use it as its identity provider. It also creates a
mixin for the up tool to fetch a token before running a test and to use
it in all requests.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>