logs: moves some loki CLI args to the config
Moves some loki configuration that previously was being passed by CLI arguments to the loki config directly in the form of ENV vars that will then be expanded to their actual values. We need this because upstream we will want to update these values without having to regenerate the templates. Furthermore, some of the values are secrets and are only known at rollout time

Signed-off-by: JoaoBraveCoding <jmarcal@redhat.com>